requirements.txt
streamlit
pandas
matplotlib
rapidfuzz
openpyxl
xlsxwriter
requests
pillow
gspread
oauth2client
pytesseract
